pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = ""
authors = [
    {name = "hhq",email = "2194482868@qq.com"}
]
readme = "README.md"
requires-python = "3.10.11"
dependencies = [
    "setuptools (>=79.0.0,<80.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "tortoise-orm[accel] (>=0.25.0,<0.26.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "aerich (>=0.8.2,<0.9.0)",
    "aiomysql (>=0.2.0,<0.3.0)",
    "uvicorn (==0.21.0)",
    "aioredis[hiredis] (>=2.0.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pydantic[email] (>=2.11.3,<3.0.0)",
    "celery (>=5.5.1,<6.0.0)",
    "eventlet (>=0.39.1,<0.40.0)",
    "redis (>=5.2.1,<6.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "langgraph (>=0.3.31,<0.4.0)",
    "langchain-community (>=0.3.22,<0.4.0)",
    "langchain-openai (>=0.3.16,<0.4.0)",
    "langgraph-swarm (>=0.0.11,<0.0.12)",
    "langgraph-supervisor (>=0.0.20,<0.0.21)",
    "tomlkit (>=0.13.2,<0.14.0)",
    "langchain-experimental (>=0.3.4,<0.4.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "e2b-code-interpreter (>=1.5.0,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "fpdf (>=1.7.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.aerich]
tortoise_orm = "settings.orm"
location = "./migrations"
src_folder = "./."
